changed file name: Projector.py to Projectors.py
